requirements.txt
# 核心依赖
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
httpx[socks]>=0.24.0
qianfan>=0.0.3
pyyaml>=6.0.0
fasttext>=0.9.2

# 测试依赖
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0